refactor: Update the fetching and updating data using concept of use-cases and services
