+++ app/server/api/batch/[id].get.ts
@@ -1,75 +1,19 @@
-import { PrismaClient } from '~/generated/prisma';
-
-const prisma = new PrismaClient();
+import { getBatchDetailsUseCase } from '~/server/use-cases/useGetBatch';
 
 export default defineEventHandler(async (event) => {
   const batchId = getRouterParam(event, 'id') as string;
 
-  if (!batchId) {
-    throw createError({
-      statusCode: 400,
-      statusMessage: 'A Batch ID is required.',
-    });
-  }
-
   try {
-
-    const batchDataFromDb = await prisma.batch.findUnique({
-      where: {
-        id: batchId,
-      },
-      select: {
-        id: true,
-        batch_number: true,
-        status: true,
-        start_date: true,
-        Intern: {     
-          select: {
-            id: true,
-            user: {
-              select: {
-                name: true,
-              },
-            },
-            hours_completed: true,
-            required_hours: true, 
-            intern_picture: true,   
-          },
-        },
-      },
-    });
-
-    if (!batchDataFromDb) {
-      throw createError({
-        statusCode: 404,
-        statusMessage: 'Batch not found.',
-      });
-    }
-
-    const response = {
-      details: {
-        id: batchDataFromDb.id,
-        batchNumber: batchDataFromDb.batch_number,
-        statusText: batchDataFromDb.status ? 'Ongoing' : 'Inactive',
-        start_date: batchDataFromDb.start_date.toISOString().split('T')[0],
-        internCount: batchDataFromDb.Intern.length,
-      },
-      interns: batchDataFromDb.Intern.map((intern) => ({
-        id: intern.id,
-        fullName: intern.user.name,
-        internPicture: intern.intern_picture,
-        hoursCompleted: intern.hours_completed,
-        requiredHours: intern.required_hours,
-      })),
-    };
-
-    return response;
+    const batchDetails = await getBatchDetailsUseCase(batchId);
     
-  } catch (error) {
+    return batchDetails;
+
+  } catch (error: any) {
     console.error('Failed to fetch batch details:', error);
+    
     throw createError({
-      statusCode: 500,
-      statusMessage: 'Could not fetch batch details.',
+      statusCode: error.statusCode || 500,
+      statusMessage: error.statusMessage || 'Could not fetch batch details.',
     });
   }
 });
