+++ app/server/api/school/index.get.ts
@@ -1,26 +1,17 @@
-import { PrismaClient } from '~/generated/prisma'
-
-const prisma = new PrismaClient()
+import { getAllSchoolsUseCase } from '~/server/use-cases/useGetAllSchool';
 
 export default defineEventHandler(async () => {
   try {
-    const interns = await prisma.intern.findMany({
-      select: {
-        school: true,
-      },
-      distinct: ['school'], 
-    })
+    const schoolNames = await getAllSchoolsUseCase();
+    
+    return schoolNames;
 
-    const schoolNames = interns
-      .map(intern => intern.school)
-      .filter(school => school); 
+  } catch (error: any) {
+    console.error('API Error: Failed to fetch schools:', error);
 
-    return schoolNames
-  } catch (error) {
-    console.error('API Error: Failed to fetch schools:', error)
     throw createError({
-      statusCode: 500,
-      statusMessage: 'Could not fetch the list of schools.',
-    })
+      statusCode: error.statusCode || 500,
+      statusMessage: error.statusMessage || 'Could not fetch the list of schools.',
+    });
   }
-})
+});
