+++ app/server/utils/prisma.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from '~/generated/prisma'
+
+const prisma = new PrismaClient()
+
+export default prisma
